requirements.txt
streamlit
pandas
plotly
openpyxl
yagmail
python-dotenv